+++ src/main/java/com/sftc/web/dao/mybatis/UserMapper.java
@@ -12,7 +12,7 @@ public interface UserMapper {
 
     List<User> selectUserByOpenid(String open_id);
 
-    User selectUserByUserId(long user_id);
+    User selectUserByUserId(int user_id);
 
     void insertOpenid(User user);
