+++ src/main/java/com/sftc/web/model/entity/Address.java
@@ -13,9 +13,9 @@ public class Address extends Object {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
     // 用户id
-    private long user_id;
+    private int user_id;
     // 用户名字
     private String name;
     // 用户电话
@@ -41,19 +41,19 @@ public class Address extends Object {
         super();
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
 
-    public Long getUser_id() {
+    public int getUser_id() {
         return user_id;
     }
 
-    public void setUser_id(long user_id) {
+    public void setUser_id(int user_id) {
         this.user_id = user_id;
     }
 
@@ -137,7 +137,7 @@ public class Address extends Object {
         this.supplementary_info = supplementary_info;
     }
 
-    public Address(Long id, long user_id, String name, String phone, String province, String city, String area, String address, String create_time) {
+    public Address(int id, int user_id, String name, String phone, String province, String city, String area, String address, String create_time) {
         this.id = id;
         this.user_id = user_id;
         this.name = name;
@@ -178,7 +178,7 @@ public class Address extends Object {
         this.setCreate_time(Long.toString(System.currentTimeMillis()));
     }
 
-    public Address(long user_id, String name, String phone, String province, String city, String area, String address, String supplementary_info, double longitude, double latitude, String create_time) {
+    public Address(int user_id, String name, String phone, String province, String city, String area, String address, String supplementary_info, double longitude, double latitude, String create_time) {
         this.user_id = user_id;
         this.name = name;
         this.phone = phone;
